added Exo player dependencies, added Exo Player instead Media Player, added audio layout for receiver

+++ app/src/main/java/com/clover/studio/exampleapp/utils/Tools.kt
@@ -308,17 +308,17 @@ object Tools {
         return UUID.randomUUID().toString().substring(0, 13)
     }
 
-    fun convertDurationMillis(time: Int): String {
-        val millis: Long = time.toLong()
+    fun convertDurationMillis(time: Long): String {
+        val millis: Long = time
         //val hour = TimeUnit.MILLISECONDS.toHours(millis)
         val minutes = TimeUnit.MILLISECONDS.toMinutes(millis) - TimeUnit.HOURS.toMinutes(
             TimeUnit.MILLISECONDS.toHours(
-                time.toLong()
+                time
             )
         )
         val seconds = TimeUnit.MILLISECONDS.toSeconds(millis) - TimeUnit.MINUTES.toSeconds(
             TimeUnit.MILLISECONDS.toMinutes(
-                time.toLong()
+                time
             )
 
         )
